apps/api/pyproject.toml
[project]
name = "ollyscale-frontend"
version = "0.1.0"
description = "ollyScale v2 Frontend - Postgres-backed observability platform"
authors = [
    {name = "Ryan Faircloth", email = "ryan@faircloth.org"}
]
license = {text = "AGPL-3.0"}
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "grpcio>=1.76.0",
    "grpcio-health-checking>=1.76.0",
    "httpx>=0.28.0",
    "opentelemetry-proto>=1.39.1",
    "opentelemetry-api>=1.28.0",
    "opentelemetry-sdk>=1.28.0",
    "opentelemetry-instrumentation>=0.49b0",
    "opentelemetry-instrumentation-logging>=0.49b0",
    "opentelemetry-instrumentation-fastapi>=0.49b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.49b0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.28.0",
    "protobuf>=6.33.4",
    "gunicorn>=24.1.1",
    "psycopg2-binary>=2.9.11",
    "uvicorn[standard]>=0.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py314"
line-length = 120
extend-exclude = ["migrations"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP", "ARG", "SIM", "TCH", "PTH", "ERA", "PL", "RUF"]
ignore = [
    "E501",     # Line too long (handled by formatter)
    "PLR0913",  # Too many arguments
    "PLR0912",  # Too many branches
    "PLR0915",  # Too many statements
    "PLR2004",  # Magic value used in comparison
    "PLR0911",  # Too many return statements
    "B008",     # Do not perform function call in argument defaults (FastAPI Depends pattern)
    "PLW0603",  # Using global statement (singleton pattern in dependencies.py)
    "SIM108",   # Use ternary operator (less readable for this case)
    "TCH001",   # Move import into type-checking block
    "TCH002",   # Move import into type-checking block
    "TCH003",   # Move import into type-checking block
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[dependency-groups]
dev = [
    "eralchemy (>=1.6.0,<2.0.0)",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=6.0.0",
    "testcontainers[postgres]>=4.14.0",
]
